# doc/openapi.yaml
openapi: 3.0.0
info:
  title: 飲食店向け簡易オーダーシステム
  version: "1.0"
servers:
  - url: "http://localhost:3000"
...
paths:
...
  /management/restaurant:
    get:
      tags:
        - management
      summary: 店舗の一覧取得 (ownerは自分の店舗のみ取得可能)
      operationId: restaurantList
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  restaurants:
                    type: array
                    items:
                      type: object
                      properties:
                        restaurant_id:
                          type: integer
                        restaurant_name:
                          type: string
                        owner_admin_id:
                          type: integer
                        restaurant_image_url:
                          type: string
        "401":
          description: Unauthorized (トークンが無効)
        "403":
          description: Permission Denied (counter, kitchenはアクセス権がない)
      parameters:
        - schema:
            type: string
          in: header
          required: true
          name: authorization
          description: /management/login で取得したトークン
      security:
        - bearerAuth: []

    post:
      tags:
        - management
      summary: 新規店舗登録
      operationId: restaurantSignUp
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  restaurant_id:
                    type: integer
                    description: 店舗のID
        "400":
          description: Bad Request (リクエストに不備がある)
        "401":
          description: Unauthorized (トークンが無効)
        "403":
          description: Permission Denied (owner, counter, kitchenはアクセス権がない)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                owner_admin_id:
                  type: integer
                  description: オーナーの管理者ID
                restaurant_name:
                  type: string
                  description: 店舗の名称
                restaurant_address:
                  type: string
                restaurant_image_url:
                  type: string
              required:
                - owner_admin_id
                - restaurant_name
      parameters:
        - schema:
            type: string
          in: header
          required: true
          name: authorization
          description: /management/login で取得したトークン
      security:
        - bearerAuth: []
  /management/restaurant/{restaurant_id}:
    get:
      tags:
        - management
      summary: 指定した店舗情報取得
      operationId: restaurantGet
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  restaurant_id:
                    type: integer
                    description: 店舗のID
                  restaurant_name:
                    type: string
                    description: 店舗の名称
                  owner_admin_id:
                    type: integer
                    description: オーナーの管理者ID
                  owner_name:
                    type: string
                    description: オーナーの名前
                  restaurant_address:
                    type: string
                  restaurant_image_url:
                    type: string
        "400":
          description: Bad Request (リクエストに不備がある)
        "401":
          description: Unauthorized (トークンが無効)
        "403":
          description: Permission Denied (owner, kitchen, counterは自分の店舗のみ取得可能)
      parameters:
        - schema:
            type: string
          in: header
          required: true
          name: authorization
          description: /management/login で取得したトークン
        - schema:
            type: integer
          in: path
          required: true
          name: restaurant_id
          description: 情報取得したい店舗のID
      security:
        - bearerAuth: []

    put:
      tags:
        - management
      summary: 店舗情報編集
      operationId: restaurantModify
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request (リクエストに不備がある)
        "401":
          description: Unauthorized (トークンが無効)
        "403":
          description: Permission Denied (ownerは自分の店舗のみ編集可。counter, kitchenはアクセス権がない)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                restaurant_name:
                  type: string
                  description: 店舗の名称
                owner_admin_id:
                  type: integer
                  description: オーナーの管理者ID
                restaurant_address:
                  type: string
                restaurant_image_url:
                  type: string
      parameters:
        - schema:
            type: string
          in: header
          required: true
          name: authorization
          description: /management/login で取得したトークン
        - schema:
            type: integer
          in: path
          required: true
          name: restaurant_id
          description: 情報編集したい店舗のID
      security:
        - bearerAuth: []

    delete:
      tags:
        - management
      summary: 店舗の削除
      operationId: restaurantDelete
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request (restaurant_idに不備がある)
        "401":
          description: Unauthorized (トークンが無効)
        "403":
          description: Permission Denied (system管理者のみアクセス可)
      parameters:
        - schema:
            type: string
          in: header
          required: true
          name: authorization
          description: /management/login で取得したトークン
        - schema:
            type: integer
          in: path
          required: true
          name: restaurant_id
          description: 削除したい店舗のID
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT